refactor: modularize Telegram library
- Applied modularization for better code organization.
- No changes to existing functionality.
